Chunks Java payID demos
